Added initial framemaker rewrite code
