+++ package.json
@@ -10,10 +10,16 @@
   "author": "Joshua Castor",
   "license": "ISC",
   "dependencies": {
+    "eslint": "^5.11.0",
     "express": "^4.15.3",
+    "greenlet": "^1.0.1",
     "puppeteer": "^1.1.1",
     "request": "^2.83.0",
     "request-promise": "^4.2.2",
+    "rollup": "^0.68.2",
     "yargs": "^8.0.2"
+  },
+  "devDependencies": {
+    "rollup-plugin-node-resolve": "^4.0.0"
   }
 }
